+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # **Upcoming release**
 
+- ...
+- #516 Autoimport Now automatically detects project dependencies and can read TOML configuration
 - #730 Match on module aliases for autoimport suggestions
 - #755 Remove dependency on `build` package being installed while running tests